refactor: move and rename "functions.preprocessor"
